Add files via upload
Login means you are already registered and your name is on the list. Your credentials are already saved with the account and you are just authenticating yourself as a returning user. Sign up, on the contrary, is an action taken by the user who identifies himself as a new user.